+++ src/Memoria.EventSourcing.Store.EntityFrameworkCore.Identity/IdentityDomainDbContext.cs
@@ -29,11 +29,13 @@ public abstract class IdentityDomainDbContext(
         modelBuilder.ApplyConfiguration(new AggregateEntityConfiguration());
         modelBuilder.ApplyConfiguration(new EventEntityConfiguration());
         modelBuilder.ApplyConfiguration(new AggregateEventEntityConfiguration());
+        modelBuilder.ApplyConfiguration(new ProjectionEntityConfiguration());
     }
 
     public DbSet<AggregateEntity> Aggregates { get; set; } = null!;
     public DbSet<EventEntity> Events { get; set; } = null!;
     public DbSet<AggregateEventEntity> AggregateEvents { get; set; } = null!;
+    public DbSet<ProjectionEntity> Projections { get; set; } = null!;
 
     public void DetachAggregate<T>(IAggregateId<T> aggregateId, T aggregate) where T : IAggregateRoot
     {
@@ -55,12 +57,12 @@ public abstract class IdentityDomainDbContext(
     {
         foreach (var entityEntry in ChangeTracker.Entries())
         {
-            if (entityEntry.Entity is not AggregateEntity aggregateEntity)
+            if (entityEntry.Entity is not ProjectionEntity projectionEntity)
             {
                 continue;
             }
 
-            if (aggregateEntity.Id == projectionId.ToStoreId())
+            if (projectionEntity.Id == projectionId.ToStoreId())
             {
                 entityEntry.State = EntityState.Detached;
             }